# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Layse Duarte API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  durationMinutes: number;
  createdAt: string;
}

export interface CreateServiceInput {
  name: string;
  description?: string | null;
  price: number;
  durationMinutes: number;
}

export interface Client {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  createdAt: string;
  totalAppointments: number;
  lastAppointment?: string | null;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  blocked: "blocked",
} as const;

export type AppointmentPaymentMethod =
  | (typeof AppointmentPaymentMethod)[keyof typeof AppointmentPaymentMethod]
  | null;

export const AppointmentPaymentMethod = {
  pix: "pix",
  cash: "cash",
  card: "card",
} as const;

export interface Appointment {
  id: number;
  clientId?: number | null;
  clientName: string;
  clientPhone?: string | null;
  serviceId: number;
  serviceName: string;
  servicePrice: number;
  date: string;
  startTime: string;
  endTime: string;
  status: AppointmentStatus;
  notes?: string | null;
  paymentMethod?: AppointmentPaymentMethod;
  createdAt: string;
}

export interface ClientWithHistory {
  id: number;
  name: string;
  phone?: string | null;
  email?: string | null;
  createdAt: string;
  appointments: Appointment[];
}

export interface CreateClientInput {
  name: string;
  phone?: string | null;
  email?: string | null;
}

export type CreateAppointmentInputStatus =
  (typeof CreateAppointmentInputStatus)[keyof typeof CreateAppointmentInputStatus];

export const CreateAppointmentInputStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  blocked: "blocked",
} as const;

export interface CreateAppointmentInput {
  clientId?: number | null;
  clientName: string;
  clientPhone?: string | null;
  serviceId: number;
  date: string;
  startTime: string;
  notes?: string | null;
  status?: CreateAppointmentInputStatus;
}

export type UpdateAppointmentInputStatus =
  (typeof UpdateAppointmentInputStatus)[keyof typeof UpdateAppointmentInputStatus];

export const UpdateAppointmentInputStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  blocked: "blocked",
} as const;

export type UpdateAppointmentInputPaymentMethod =
  | (typeof UpdateAppointmentInputPaymentMethod)[keyof typeof UpdateAppointmentInputPaymentMethod]
  | null;

export const UpdateAppointmentInputPaymentMethod = {
  pix: "pix",
  cash: "cash",
  card: "card",
} as const;

export interface UpdateAppointmentInput {
  clientId?: number | null;
  clientName?: string;
  clientPhone?: string | null;
  serviceId?: number;
  date?: string;
  startTime?: string;
  notes?: string | null;
  status?: UpdateAppointmentInputStatus;
  paymentMethod?: UpdateAppointmentInputPaymentMethod;
}

export type UpdateStatusInputStatus =
  (typeof UpdateStatusInputStatus)[keyof typeof UpdateStatusInputStatus];

export const UpdateStatusInputStatus = {
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
  blocked: "blocked",
} as const;

export type UpdateStatusInputPaymentMethod =
  | (typeof UpdateStatusInputPaymentMethod)[keyof typeof UpdateStatusInputPaymentMethod]
  | null;

export const UpdateStatusInputPaymentMethod = {
  pix: "pix",
  cash: "cash",
  card: "card",
} as const;

export interface UpdateStatusInput {
  status: UpdateStatusInputStatus;
  paymentMethod?: UpdateStatusInputPaymentMethod;
}

export type IncomePaymentMethod =
  (typeof IncomePaymentMethod)[keyof typeof IncomePaymentMethod];

export const IncomePaymentMethod = {
  pix: "pix",
  cash: "cash",
  card: "card",
} as const;

export interface Income {
  id: number;
  appointmentId?: number | null;
  serviceName: string;
  amount: number;
  paymentMethod: IncomePaymentMethod;
  date: string;
  clientName?: string | null;
  createdAt: string;
}

export type CreateIncomeInputPaymentMethod =
  (typeof CreateIncomeInputPaymentMethod)[keyof typeof CreateIncomeInputPaymentMethod];

export const CreateIncomeInputPaymentMethod = {
  pix: "pix",
  cash: "cash",
  card: "card",
} as const;

export interface CreateIncomeInput {
  appointmentId?: number | null;
  serviceName: string;
  amount: number;
  paymentMethod: CreateIncomeInputPaymentMethod;
  date: string;
  clientName?: string | null;
}

export interface Expense {
  id: number;
  description: string;
  amount: number;
  category: string;
  date: string;
  createdAt: string;
}

export interface CreateExpenseInput {
  description: string;
  amount: number;
  category: string;
  date: string;
}

export type FinancialSummaryIncomeByPaymentMethod = {
  pix: number;
  cash: number;
  card: number;
};

export type FinancialSummaryExpensesByCategoryItem = {
  category: string;
  total: number;
};

export type FinancialSummaryDailyRevenueItem = {
  date: string;
  amount: number;
};

export interface FinancialSummary {
  totalIncome: number;
  totalExpenses: number;
  netProfit: number;
  incomeByPaymentMethod: FinancialSummaryIncomeByPaymentMethod;
  expensesByCategory: FinancialSummaryExpensesByCategoryItem[];
  dailyRevenue: FinancialSummaryDailyRevenueItem[];
}

export type DashboardDataTopServicesItem = {
  serviceName: string;
  count: number;
  revenue: number;
};

export type DashboardDataMonthlyRevenueItem = {
  month: string;
  amount: number;
};

export interface DashboardData {
  todayRevenue: number;
  weekRevenue: number;
  monthRevenue: number;
  todayAppointments: number;
  weekAppointments: number;
  monthAppointments: number;
  averageTicket: number;
  upcomingAppointments: Appointment[];
  topServices: DashboardDataTopServicesItem[];
  monthlyRevenue: DashboardDataMonthlyRevenueItem[];
}

export interface Settings {
  studioName: string;
  primaryColor: string;
  welcomeMessage?: string | null;
  bookingMessage?: string | null;
}

export interface UpdateSettingsInput {
  studioName?: string;
  primaryColor?: string;
  welcomeMessage?: string | null;
  bookingMessage?: string | null;
}

export interface WorkingHours {
  id: number;
  dayOfWeek: number;
  dayName: string;
  isOpen: boolean;
  openTime?: string | null;
  closeTime?: string | null;
  breakStart?: string | null;
  breakEnd?: string | null;
}

export interface UpdateWorkingHoursInput {
  dayOfWeek: number;
  isOpen: boolean;
  openTime?: string | null;
  closeTime?: string | null;
  breakStart?: string | null;
  breakEnd?: string | null;
}

export interface PublicBookingInput {
  clientName: string;
  clientPhone?: string | null;
  clientEmail?: string | null;
  serviceId: number;
  date: string;
  startTime: string;
  notes?: string | null;
}

export type ListAppointmentsParams = {
  startDate?: string;
  endDate?: string;
  status?: string;
};

export type ListIncomesParams = {
  startDate?: string;
  endDate?: string;
};

export type ListExpensesParams = {
  startDate?: string;
  endDate?: string;
};

export type GetFinancialSummaryParams = {
  startDate?: string;
  endDate?: string;
};

export type GetAvailabilityParams = {
  date: string;
  serviceId: number;
};