# lib/api-zod/src/generated/types/createAppointmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Layse Duarte API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAppointmentInputStatus } from "./createAppointmentInputStatus";

export interface CreateAppointmentInput {
  clientId?: number | null;
  clientName: string;
  clientPhone?: string | null;
  serviceId: number;
  date: string;
  startTime: string;
  notes?: string | null;
  status?: CreateAppointmentInputStatus;
}