# lib/api-zod/src/generated/types/createExpenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Studio Layse Duarte API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExpenseInput {
  description: string;
  amount: number;
  category: string;
  date: string;
}